Rename obtain-token.js to obtain-access-token.js
To better represent content of file.
